Rename climber folder/package to stilts
