+++ src/main/java/frc/team4909/robot/Robot.java
@@ -3,7 +3,6 @@ package frc.team4909.robot;
 import edu.wpi.first.wpilibj.Compressor;
 import edu.wpi.first.wpilibj.PowerDistributionPanel;
 import edu.wpi.first.wpilibj.TimedRobot;
-import edu.wpi.first.wpilibj.command.Command;
 import edu.wpi.first.wpilibj.command.Scheduler;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj.DriverStation;
@@ -12,8 +11,8 @@ import frc.team4909.robot.sensors.LidarLitePWM;
 import frc.team4909.robot.sensors.Stream;
 import frc.team4909.robot.subsystems.StiltWheel.StiltWheelSubsystem;
 import frc.team4909.robot.subsystems.StiltWheel.commands.MoveStiltWheels;
-import frc.team4909.robot.subsystems.climber.ClimberSubsystem;
-import frc.team4909.robot.subsystems.climber.commands.*;
+import frc.team4909.robot.subsystems.stilts.ClimberSubsystem;
+import frc.team4909.robot.subsystems.stilts.commands.*;
 import frc.team4909.robot.subsystems.drivetrain.DriveTrainSubsystem;
 import frc.team4909.robot.subsystems.drivetrain.commands.InvertDriveDirection;
 import frc.team4909.robot.subsystems.drivetrain.commands.TogglePreciseMode;
