+++ src/main/java/frc/team4909/robot/subsystems/stilts/commands/MoveStiltsOnly.java
@@ -1,4 +1,4 @@
-package frc.team4909.robot.subsystems.climber.commands;
+package frc.team4909.robot.subsystems.stilts.commands;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.team4909.robot.Robot;
